feat(Settings): add settings and merge from master
